Conflicts:
	src/mainboard/asus/m5a88-v/devicetree.cb
	src/mainboard/asus/m5a88-v/mb_sysconf.h
	src/mainboard/asus/m5a88-v/platform_cfg.h
	src/mainboard/asus/m5a88-v/reset.c
	src/mainboard/asus/m5a88-v/resourcemap.c
	src/mainboard/asus/m5a88-v/romstage.c

+++ src/mainboard/tyan/s8226/mptable.c
@@ -14,7 +14,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301 USA
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
 
@@ -33,7 +33,7 @@ extern u8 bus_sp5100[2];
 extern u32 bus_type[256];
 extern u32 sbdn_sr5650;
 extern u32 sbdn_sp5100;
-extern u8 bus_isa;
+extern int bus_isa;
 
 
 static void *smp_write_config_table(void *v)
